Add std::rcu::unique_ptr<T> per my email discussion.
std::rcu::make_unique<T>(...) returns a std::rcu::unique_ptr<T> object
associated with a heap-allocated T object (the "controlled object").
When the unique_ptr object goes out of scope or loses its value for
whatever reason (e.g. reassignment to a different value), the
implementation defers a call to "delete controlled_object".

This is NOT EVEN CLOSE to David Hollman's rcu_ptr<T>, which is more
of a way of splitting up data and metadata. I'll work on that next.