Squash merge of branch `ajo-cell`
Add an implementation of P0561R0 (Romer and Hunter's cell<T>).

I claim that this class should never be used, because it holds
RCU reader locks for arbitrarily long periods of time. Creating
a global variable of snapshot_ptr type (and initializing it with
a snapshot) will disable RCU reclamation forever.