Add the "SWMRList" example from Folly, and test it.
As of https://github.com/facebook/folly/commit/ed8c80a0, the Folly repo's
version of `SWMRList<T>` (a single-writer, multi-reader set implemented as
a sorted linked list) has bugs. But I believe I've fixed them all in this
version. There was really only one significant bug: the omission of setting
a removed node's "next" pointer to `nullptr`. (And then there's a bit of a
trick: you really do have to remove the node first and *then* set its "next"
pointer to null, because if you do the null assignment first, you'll truncate
the list, which means that some readers will see a truncated list and return
the wrong answer. I've added comments to try to explain this to future-me.)