Reimplement rcu::cell<T> in a safe way, with reference counting.
Based on an idea from Paul McKenney. This implementation avoids the use
of `rcu_reader_lock()`, which eliminates the possibility of long-lived
reader locks. Instead, `cell<T>` becomes basically a synonym for
`std::atomic_shared_ptr<T>`, but with a custom deleter that calls
`retire` instead of `delete` when the last reference is dropped.

At this point it's simple enough that I would probably argue that we
don't need to standardize this thing; is it really just `shared_ptr`
with a custom deleter? (And for the vast majority of users, I'd guess
that the custom deleter isn't needed either.)